requirements.txt
torch==2.0.0
h5py~=2.10.0
tqdm~=4.59.0
numpy~=1.20.1
torchvision~=0.15.0
pillow~=8.2.0